refactor: stac test compares to types.json, moved metadata test dir
